% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{assignToCluster}
\alias{assignToCluster}
\title{A Cluster Assignment function}
\usage{
assignToCluster(trainData, clusters, normalized, testData)
}
\arguments{
\item{trainData}{Dataset to be considered for clustering using Kmeans}

\item{clusters}{Denotes number of clusters. Takes input range 1-20. Defaults to 2}

\item{normalized}{Takes input TRUE/FALSE to normalize the dataset. Defaults to FALSE}

\item{testData}{Dataset to be considered to assign its data points on clusters}
}
\description{
This function allows you to assign the test data points to the clusters that are formed by train dataset
}
\examples{
assignToCluster(trainData = dataset1, clusters = 2, normalized = FALSE, testData = dataset2)
}
\keyword{OptimumClusters}
