% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{assignToClusterGMM}
\alias{assignToClusterGMM}
\title{A Cluster Assignment function}
\usage{
assignToClusterGMM(trainData, normalized, testData)
}
\arguments{
\item{trainData}{Dataset to be considered for clustering}

\item{normalized}{Takes input TRUE/FALSE to normalize the dataset. Defaults to FALSE}

\item{testData}{Dataset to be considered to assign its data points on clusters}
}
\description{
This function allows you to assign the test data points to the clusters that are formed by train dataset using GMM.
}
\examples{
assignToClusterGMM(trainData = dataset1, normalized = FALSE, testData = dataset2)
}
\keyword{OptimumClusters}
