% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{classifyKNN}
\alias{classifyKNN}
\title{A Classification function}
\usage{
classifyKNN(trainData, clusters, normalized, testData, nearPoints)
}
\arguments{
\item{trainData}{Dataset to be considered for clustering using Kmeans}

\item{clusters}{Denotes number of clusters. Takes input range 1-15. Defaults to 3}

\item{normalized}{Takes input TRUE/FALSE to normalize the dataset. Defaults to FALSE}

\item{testData}{Dataset to be considered to assign its data points on clusters}

\item{nearPoints}{Number of neighbor train data points to be considered}
}
\description{
This function allows you to predict the gender of the test data points.
}
\examples{
classifyKNN(trainData = dataset1, clusters = 3, normalized = FALSE, testData = dataset2, nearPoints = 3)
}
\keyword{classifyKNN}
