% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{clusterPlot3d}
\alias{clusterPlot3d}
\title{A 3D Cluster Plot function}
\usage{
clusterPlot3d(trainData, clusters, normalized)
}
\arguments{
\item{trainData}{Dataset to be passed for clustering}

\item{clusters}{Denotes number of clusters. Takes input range 1-20. Defaults to 2}

\item{normalized}{Takes input TRUE/FALSE to normalize the dataset. Defaults to FALSE}
}
\description{
This function allows you to plot the data points in clusters using Kmeans
}
\examples{
clusterPlot3d(dataset, clusters = 3, normalized = FALSE)
}
\keyword{ClusterPlot}
