% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{clusterPlot3dGMM}
\alias{clusterPlot3dGMM}
\title{A 3D Cluster Plot function}
\usage{
clusterPlot3dGMM(trainData, normalized)
}
\arguments{
\item{trainData}{Dataset to be passed for clustering}

\item{normalized}{Takes input TRUE/FALSE to normalize the dataset. Defaults to FALSE}
}
\description{
This function allows you to plot the data points in clusters using GMM
}
\examples{
clusterPlot3d(dataset, clusters = 3, normalized = FALSE)
}
\keyword{ClusterPlot}
