% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{findGender}
\alias{findGender}
\title{A findGender function}
\usage{
findGender(trainDataWithCluster, testDataWithCluster, nearPoints,
  originalTrainData, originalTestData)
}
\arguments{
\item{nearPoints}{Number of neighbor train data points to be considered. Defaults to sqrt of no. of observations in a train dataset}

\item{trainData}{Dataset to be considered for clustering using Kmeans}

\item{clusters}{Denotes number of clusters. Takes input range 1-15. Defaults to 3}

\item{testData}{Dataset to be considered to assign its data points on clusters}

\item{normalized}{Takes input TRUE/FALSE to normalize the dataset. Defaults to FALSE}
}
\description{
This function allows you to find nearest n points in a cluster to the test data point to predict the Gender
}
\examples{
findGender(trainDataWithCluster, testDataWithCluster, nearPoints)
}
\keyword{gender}
