% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{mapDataToCluster}
\alias{mapDataToCluster}
\title{A Map data points to Clusters function}
\usage{
mapDataToCluster(testData, centers)
}
\arguments{
\item{testData}{Dataset to be considered to assign its data points on clusters}

\item{centers}{Takes value. Defaults to 2}
}
\description{
This function allows you to find appropriate cluster for the test data points using distance from center of the clusters
}
\examples{
mapDataToCluster(testData = dataset, centers = 2 )
}
\keyword{mapDataToClusters}
