% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{normalizeData}
\alias{normalizeData}
\title{A Normalization function}
\usage{
normalizeData(trainData)
}
\arguments{
\item{trainData}{Dataset to be normalized}
}
\description{
This function allows you to normalize the variables of the dataset
}
\examples{
normalizeData(dataset)
}
\keyword{normalize}
