% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{optimumClusters}
\alias{optimumClusters}
\title{A Optimum Clusters function}
\usage{
optimumClusters(trainData, normalized)
}
\arguments{
\item{trainData}{Dataset to be considered}

\item{normalized}{Takes input TRUE/FALSE to normalize the dataset. Defaults to FALSE}
}
\description{
This function allows you to find the optimum number of clusters
}
\examples{
optimumClusters(dataset, normalized = FALSE)
}
\keyword{OptimumClusters}
