% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{plotClusterDatasetFromPredicted}
\alias{plotClusterDatasetFromPredicted}
\title{A 3D Cluster Plot function using Kmeans}
\usage{
plotClusterDatasetFromPredicted(trainDataWithCluster)
}
\arguments{
\item{trainDataWithCluster}{Test dataset with assigned clusters}
}
\description{
This function allows you to plot the test data points in the clusters, formed by train dataset
}
\examples{
plotClusterDatasetFromPredicted(dataset)
}
\keyword{plotCluster}
