% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runClusterApplication}
\alias{runClusterApplication}
\title{A 3D Cluster Plot function using Kmeans}
\usage{
runClusterApplication()
}
\description{
This function allows you to plot the test data points in the clusters, formed by train dataset
}
\examples{
runClusterApplication()
}
\keyword{plotCluster}
