% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{scatter3dPerson}
\alias{scatter3dPerson}
\title{A 3D Scatterplot function}
\usage{
scatter3dPerson(trainData)
}
\arguments{
\item{trainData}{Dataset to be ploted}
}
\description{
This function allows you to plot the data points in 3D
}
\examples{
scatter3dPerson(dataset)
}
\keyword{scatter3d}
